+++ lib/id_card_formatter.dart
@@ -1,3 +1,3 @@
 part 'double_format.dart';
 part 'id_format.dart';
-
+part 'generic_string_format.dart';