+++ test/id_card_formatter_test.dart
@@ -10,5 +10,14 @@ void main() {
     expect(ncard, '1234 5678 9012 3456');
     final voter = "T123456789012".tzVoterFormat();
     expect(voter, 'T-1234-5678-901-2');
+    final genearicVoter =
+        "T123456789012".genericFormat(format: 'X-XXXX-XXXX-XXX-X');
+    expect(genearicVoter, 'T-1234-5678-901-2');
+    final genearicNida =
+        "12345678123451234512".genericFormat(format: 'XXXXXXXX-XXXXX-XXXXX-XX');
+    expect(genearicNida, '12345678-12345-12345-12');
+    final genearicNCard =
+        "1234567890123456".genericFormat(format: 'XXXX XXXX XXXX XXXX');
+    expect(genearicNCard, '1234 5678 9012 3456');
   });
 }